+++ packages/extensions/core/src/lib/plugins/loaders/openapi-loader.ts
@@ -1,9 +1,6 @@
 import { PipelinePlugin } from "../../pipeline/common";
-
 import { Channel } from "../../message";
-import { parse as ParseLiterateYaml } from "@autorest/common";
-import { CloneAst, DataHandle, DataSink, DataSource, QuickDataSource, StringifyAst } from "@azure-tools/datastore";
-import { identitySourceMapping } from "@autorest/common";
+import { DataHandle, DataSink, DataSource, QuickDataSource } from "@azure-tools/datastore";
 import { crawlReferences } from "../ref-crawling";
 import { AutorestContext } from "../../context";
 import { checkSyntaxFromData } from "./common";
@@ -15,7 +12,7 @@ interface OpenAPI3Spec {
   components?: { schemas?: object };
 }
 
-export async function LoadLiterateOpenAPIs(
+export async function loadOpenAPIFiles(
   config: AutorestContext,
   inputScope: DataSource,
   inputFileUris: Array<string>,
@@ -24,7 +21,7 @@ export async function LoadLiterateOpenAPIs(
   const rawOpenApis: Array<DataHandle> = [];
   for (const inputFileUri of inputFileUris) {
     // read literate Swagger
-    const pluginInput = await LoadLiterateOpenAPI(config, inputScope, inputFileUri, sink);
+    const pluginInput = await loadOpenAPIFile(config, inputScope, inputFileUri, sink);
     if (pluginInput) {
       rawOpenApis.push(pluginInput);
     }
@@ -32,25 +29,21 @@ export async function LoadLiterateOpenAPIs(
   return rawOpenApis;
 }
 
-export async function LoadLiterateOpenAPI(
+export async function loadOpenAPIFile(
   config: AutorestContext,
   inputScope: DataSource,
   inputFileUri: string,
   sink: DataSink,
 ): Promise<DataHandle | null> {
-  const handle = await inputScope.ReadStrict(inputFileUri);
+  const handle = await inputScope.readStrict(inputFileUri);
   await checkSyntaxFromData(inputFileUri, handle, config);
-  const data = await ParseLiterateYaml(config, handle, sink);
-  if (!isOpenAPI3Spec(await data.ReadObject<OpenAPI3Spec>())) {
+  // const data = await ParseLiterateYaml(config, handle, sink);
+  if (!isOpenAPI3Spec(await handle.readObject<OpenAPI3Spec>())) {
     return null;
     // TODO: Should we throw or send an error message?
   }
   config.Message({ Channel: Channel.Verbose, Text: `Reading OpenAPI 3.0 file ${inputFileUri}` });
-
-  const ast = CloneAst(await data.ReadYamlAst());
-  const mapping = identitySourceMapping(data.key, ast);
-
-  return sink.WriteData(handle.Description, StringifyAst(ast), [inputFileUri], "openapi-document", mapping, [data]);
+  return sink.writeData(handle.description, await handle.readData(), [inputFileUri], "openapi-document");
 }
 
 /**
@@ -65,8 +58,8 @@ function isOpenAPI3Spec(specObject: OpenAPI3Spec): boolean {
 export function createOpenApiLoaderPlugin(): PipelinePlugin {
   return async (config, input, sink) => {
     const inputs = config.config.inputFileUris;
-    const openapis = await LoadLiterateOpenAPIs(config, input, inputs, sink);
-    let result: Array<DataHandle> = [];
+    const openapis = await loadOpenAPIFiles(config, input, inputs, sink);
+    let result: DataHandle[] = [];
     if (openapis.length === inputs.length) {
       result = await crawlReferences(config, input, openapis, sink);
     }