+++ packages/extensions/core/src/lib/plugins/yaml-and-json.ts
@@ -4,17 +4,17 @@ import { createPerFilePlugin, PipelinePlugin } from "../pipeline/common";
 /* @internal */
 export function createYamlToJsonPlugin(): PipelinePlugin {
   return createPerFilePlugin(async () => async (fileIn, sink) => {
-    let ast = await fileIn.ReadYamlAst();
+    let ast = await fileIn.readYamlAst();
     ast = ConvertYaml2Jsonx(ast);
-    return sink.WriteData(fileIn.Description, StringifyAst(ast), fileIn.identity);
+    return sink.writeData(fileIn.description, StringifyAst(ast), fileIn.identity);
   });
 }
 
 /* @internal */
 export function createJsonToYamlPlugin(): PipelinePlugin {
   return createPerFilePlugin(async () => async (fileIn, sink) => {
-    let ast = await fileIn.ReadYamlAst();
+    let ast = await fileIn.readYamlAst();
     ast = ConvertJsonx2Yaml(ast);
-    return sink.WriteData(fileIn.Description, StringifyAst(ast), fileIn.identity);
+    return sink.writeData(fileIn.description, StringifyAst(ast), fileIn.identity);
   });
 }