+++ packages/libs/configuration/src/autorest-normalized-configuration.ts
@@ -37,6 +37,11 @@ export interface AutorestNormalizedConfiguration extends AutorestRawConfiguratio
   "stats"?: boolean;
 
   /**
+   * Start the autorest.interactive plugin and cache traffic between extensions.
+   */
+  "interactive"?: boolean;
+
+  /**
    * Skip the semantic validation plugin.
    */
   "skip-semantics-validation"?: boolean;
@@ -56,6 +61,11 @@ export interface AutorestNormalizedConfiguration extends AutorestRawConfiguratio
   "batch"?: boolean;
   "resource-schema-batch"?: any;
   "perform-load"?: any;
+
+  /**
+   * End of line for generated files.
+   */
+  "eol"?: "default" | "lf" | "crlf";
 
   /**
    * Feature flags. Those flags enable/disable certain features
@@ -99,6 +109,21 @@ export interface AutorestNormalizedConfiguration extends AutorestRawConfiguratio
 
   "resolved-directive"?: any;
   "debugger"?: any;
+
+  /**
+   * Skip sourcemap generation. To speedup.
+   */
+  "skip-sourcemap"?: boolean;
+
+  /**
+   * Force updating the version of core even if there is a local version satisfying the requirement.
+   */
+  "force"?: boolean;
+
+  /**
+   * Configure max memory allowed for autorest process(s)
+   */
+  "memory"?: string;
 
   "github-auth-token"?: string;