+++ packages/libs/datastore/src/data-store/data-handle.ts
@@ -1,22 +1,19 @@
-import { Delay, Lazy, LazyPromise } from "@azure-tools/tasks";
 import { MappedPosition, Position, RawSourceMap, SourceMapConsumer } from "source-map";
 import { promises as fs } from "fs";
 import { ParseToAst as parseAst, YAMLNode, parseYaml, ParseNode } from "../yaml";
-
-export interface Metadata {
-  lineIndices: Lazy<Array<number>>;
-  sourceMap: LazyPromise<RawSourceMap>;
-
-  // inputSourceMap: LazyPromise<RawSourceMap>;
-}
+import { getLineIndices } from "../parsing/text-utility";
 
 export interface Data {
+  status: "loaded" | "unloaded";
   name: string;
   artifactType: string;
-  metadata: Metadata;
-  identity: Array<string>;
+  identity: string[];
+
+  lineIndices?: number[];
+  sourceMap: RawSourceMap | undefined;
 
   writeToDisk?: Promise<void>;
+  writeSourceMapToDisk?: Promise<void>;
   cached?: string;
   cachedAst?: YAMLNode;
   cachedObject?: any;
@@ -24,6 +21,8 @@ export interface Data {
 }
 
 export class DataHandle {
+  private unloadTimer: NodeJS.Timer | undefined;
+
   /**
    * @param autoUnload If the data unhandle should automatically unload files after they are not used for a while.
    */
@@ -31,7 +30,7 @@ export class DataHandle {
     // start the clock once this has been created.
     // this ensures that the data cache will be flushed if not
     // used in a reasonable amount of time
-    this.onTimer();
+    this.resetUnload();
   }
 
   public async serialize() {
@@ -45,38 +44,39 @@ export class DataHandle {
     });
   }
 
-  private onTimer() {
-    this.checkIfNeedToUnload().catch((e) => {
-      // eslint-disable-next-line no-console
-      console.error("Error while verifing DataHandle cache status", e);
-    });
-  }
+  private resetUnload() {
+    if (this.unloadTimer) {
+      clearTimeout(this.unloadTimer);
+    }
 
-  private async checkIfNeedToUnload() {
     if (!this.autoUnload) {
       return;
     }
 
-    await Delay(3000);
-    if (this.item.accessed) {
-      // it's been cached. start the timer!
-      this.onTimer();
-      // clear the accessed flag before we go.
-      this.item.accessed = false;
-      return;
-    }
-    // wasn't actually used since the delay. let's dump it.
-    // console.log(`flushing ${this.item.name}`);
-    // wait to make sure it's finished writing to disk tho'
-    // await this.item.writingToDisk;
+    setTimeout(() => {
+      if (this.item.accessed) {
+        this.item.accessed = false;
+        this.resetUnload();
+      } else {
+        this.unload();
+      }
+    }, 3000);
+  }
+
+  private unload() {
     if (!this.item.writeToDisk) {
       // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
       this.item.writeToDisk = fs.writeFile(this.item.name, this.item.cached!);
     }
+    if (this.item.sourceMap && !this.item.writeSourceMapToDisk) {
+      this.item.writeSourceMapToDisk = fs.writeFile(`${this.item.name}.map`, JSON.stringify(this.item.sourceMap));
+    }
     // clear the caches.
+    this.item.status = "unloaded";
     this.item.cached = undefined;
     this.item.cachedObject = undefined;
     this.item.cachedAst = undefined;
+    this.item.sourceMap = undefined;
   }
 
   public get originalDirectory() {
@@ -101,14 +101,14 @@ export class DataHandle {
     if (this.item.cached === undefined) {
       // make sure the write-to-disk is finished.
       await this.item.writeToDisk;
-
       if (nocache) {
         return await fs.readFile(this.item.name, "utf8");
       } else {
         this.item.cached = await fs.readFile(this.item.name, "utf8");
+        this.item.status = "loaded";
 
         // start the timer again.
-        this.onTimer();
+        this.resetUnload();
       }
     }
 
@@ -118,7 +118,6 @@ export class DataHandle {
   public async readObjectFast<T>(): Promise<T> {
     // we're going to use the data, so let's not let it expire.
     this.item.accessed = true;
-
     return this.item.cachedObject || (this.item.cachedObject = parseYaml(await this.readData()));
   }
 
@@ -135,10 +134,6 @@ export class DataHandle {
     this.item.accessed = true;
     // return the cachedAst or get it, then return it.
     return this.item.cachedAst || (this.item.cachedAst = parseAst(await this.readData()));
-  }
-
-  public get metadata(): Metadata {
-    return this.item.metadata;
   }
 
   public get artifactType(): string {
@@ -159,13 +154,38 @@ export class DataHandle {
   }
 
   public async blame(position: Position): Promise<Array<MappedPosition>> {
-    const metadata = this.metadata;
-    const consumer = new SourceMapConsumer(await metadata.sourceMap);
+    await this.readData();
+    const sourceMap = await this.getSourceMap();
+    if (!sourceMap) {
+      return [];
+    }
+    const consumer = await new SourceMapConsumer(sourceMap);
     const mappedPosition = consumer.originalPositionFor(position);
     if (mappedPosition.line === null) {
       return [];
     }
-    return [mappedPosition];
+    return [mappedPosition as any];
+  }
+
+  public async lineIndices() {
+    if (!this.item.lineIndices) {
+      this.item.lineIndices = getLineIndices(await this.readData());
+    }
+
+    return this.item.lineIndices;
+  }
+
+  public async getSourceMap() {
+    if (!this.item.sourceMap) {
+      try {
+        const content = await fs.readFile(`${this.item.name}.map`, "utf8");
+        this.item.sourceMap = JSON.parse(content.toString());
+      } catch {
+        return undefined;
+      }
+    }
+
+    return this.item.sourceMap;
   }
 
   /**
@@ -188,7 +208,6 @@ export class DataHandle {
   public get Description(): string {
     return this.description;
   }
-
   /**
    * @deprecated use @see readData
    */