+++ packages/libs/extension/src/package-manager.ts
@@ -1,3 +1,6 @@
+import { writeFile, access } from "fs/promises";
+import { join } from "path";
+
 export type PackageManagerType = "yarn" | "npm";
 
 export interface InstallOptions {
@@ -10,4 +13,27 @@ export interface PackageManager {
   clean(directory: string): Promise<void>;
 
   getPackageVersions(directory: string, packageName: string): Promise<any[]>;
+}
+
+/**
+ * Ensure there is a pacakge.json in the install directory.
+ * This is to ensure that yarn add will not look for a parent package.json and install in the parent folder instead.
+ * @param directory Directory where the package will be installed.
+ */
+export async function ensurePackageJsonExists(directory: string) {
+  const filename = join(directory, "package.json");
+  if (await exists(filename)) {
+    return;
+  }
+
+  await writeFile(filename, "{}");
+}
+
+async function exists(filename: string) {
+  try {
+    await access(filename);
+    return true;
+  } catch {
+    return false;
+  }
 }