+++ packages/libs/oai2-to-oai3/src/converter.ts
@@ -6,7 +6,6 @@ import {
   OpenAPI2ResponseHeader,
   OpenAPI2Operation,
   OpenAPI2OperationResponse,
-  OpenAPI2Parameter,
   OpenAPI2BodyParameter,
   OpenApi2FormDataParameter,
 } from "./oai2";