+++ packages/libs/configuration/CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log - @autorest/configuration
 
-This log was last generated on Thu, 20 May 2021 16:41:13 GMT and should not be manually modified.
+This log was last generated on Thu, 03 Jun 2021 22:37:55 GMT and should not be manually modified.
+
+## 1.5.0
+Thu, 03 Jun 2021 22:37:55 GMT
+
+### Minor changes
+
+- Add support for adl
+
+## 1.4.1
+Wed, 26 May 2021 18:31:17 GMT
+
+### Patches
+
+- **Fix** issue when using --use: it would try to use yarn but couldn't find yarn/cli.js location
 
 ## 1.4.0
 Thu, 20 May 2021 16:41:13 GMT